requirements.txt
boto3==1.12.11
botocore==1.15.11
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.10
django-crispy-forms==1.8.1
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.7
jmespath==0.9.5
joblib==0.14.1
Markdown==3.2.1
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
plotly==4.5.0
psycopg2==2.8.4
python-dateutil==2.8.0
pytz==2019.3
requests==2.20.1
retrying==1.3.3
scikit-learn==0.22
scipy==1.4.1
six==1.11.0
sqlparse==0.3.1
urllib3==1.24.3
whitenoise==5.0.1
xgboost==0.90
